Deplacement des étoiles dans fichiers c et h, Etoiles essai avec jaune {incorrect}

+++ OzmaWars/jni/src/star.h
@@ -4,10 +4,10 @@
 #include "SDL.h"
 
 typedef struct {
-    int r;
-    int g;
-    int b;
-    int a;
+    Uint8 r;
+    Uint8 g;
+    Uint8 b;
+    Uint8 a;
 } RGBA;
 
 typedef struct {
